hybridswap: fix unsigned underflow in zram_watermark_ok
diff_buffers = high_mem_watermark - usable_mem is computed as unsigned
int before widening to long long. When usable memory exceeds the high
watermark, the subtraction wraps to a huge positive value, so
wm = min(scale, scale - diff_buffers) collapses to a large negative and
cur_scale > wm is always true. This makes zram_watermark_ok() wrongly
report 'not ok' and wake swapd even when buffers are plentiful.

Cast both operands to long long before subtracting so the intended
negative diff (plentiful memory) is preserved and min() selects the
plain scale value.

+++ drivers/moto_swap/hybridswap/hybridswap_swapd.c
@@ -1254,8 +1254,8 @@ bool zram_watermark_ok(void)
 	unsigned long zram_used = hybridswap_fetch_zram_used_pages();
 	const unsigned int percent_constant = 100;
 
-	diff_buffers = fetch_high_mem_watermark_value() -
-		system_cur_usable_mem();
+	diff_buffers = (long long)fetch_high_mem_watermark_value() -
+		(long long)system_cur_usable_mem();
 	diff_buffers *= SZ_1M / PAGE_SIZE;
 	diff_buffers *= fetch_compress_scale_value() / 10;
 	diff_buffers = diff_buffers * percent_constant / fetch_nr_zram_total();
